Read from Cargo.toml whether a crate is published
